requirements.txt
asgiref==3.8.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
constantly==23.10.4
cryptography==42.0.4
dj-database-url==2.1.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-jazzmin==2.6.1
django-shortcuts==1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
Django==5.0.7
drf-spectacular==0.26.5
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
msgpack==1.0.7
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.5.0
treblle==0.14
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.5.0
zope.interface==6.1